freeing functions separated and added to Makefile and .h file
