+++ minishell.h
@@ -6,7 +6,7 @@
 /*   By: amakela <amakela@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/24 19:04:40 by amakela           #+#    #+#             */
-/*   Updated: 2024/04/24 19:32:29 by amakela          ###   ########.fr       */
+/*   Updated: 2024/04/28 17:54:44 by amakela          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,23 @@
 #include <unistd.h>
 #include <curses.h>
 #include <string.h>
+#include <stdbool.h>
 #include <readline/readline.h>
 #include <readline/history.h>
 #include "libft/libft.h"
 
+typedef	struct node
+{
+	char		**redirs;
+	char		*cmd;
+	int			fd_in;
+	int			fd_out;
+	struct node	*next;
+} t_node;
 
+
+void	free_list(t_node **processes);
+void	free_node(t_node *node);
+void	free_str_array(char **array);
 
 #endif
